Removing "Login" support.
There are no events launched by the PluggableAuthService plugin for
Login/Logout so there is nothing to hook into. Functionality can be
implemented with Products.PlonePAS as events are notified by that
product. Since this is intended as a plain Zope package that is not
available. Login/Logout can be inferred from the "Traversal" logs.

Fixing the LogErrors decorator; the return values of the functions being
decorated weren't being returned.

Modifying ``BaseEvent.__call__`` signature to support just the ``event``
argument in case subscribing to an event class instead of a context/event pair.

Adding ``javascript`` to the list of ignored content types for ``Traversal``.